finishing redis impl, making bolt and redis backwards compatible
